Added persistence of raids via node-persist - persisted data is stored with a key per-raid that gets updated on-the-fly with current raid data as its state changes Gyms stored in raids just as gym id - cut down raid size more for persistence
