+++ package.json
@@ -4,12 +4,13 @@
   "description": "Professor Willow Discord bot for Pokemon Go",
   "main": "index.js",
   "dependencies": {
-    "discord.js": "^11.1.0",
+    "discord.js": "^11.2.1",
     "discord.js-commando": "^0.9.0",
     "he": "^1.1.1",
     "lunr": "^2.1.3",
     "moment": "^2.18.1",
-    "node-cleanup": "^2.1.2"
+    "node-cleanup": "^2.1.2",
+    "node-persist": "^2.1.0"
   },
   "devDependencies": {},
   "scripts": {
